+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/lecture_detail/DayTimePickerSheet.kt
@@ -14,6 +14,7 @@ import androidx.compose.ui.unit.sp
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.compose.*
 import com.wafflestudio.snutt2.lib.*
+import com.wafflestudio.snutt2.lib.data.SNUTTStringUtils.toFormattedTimeString
 import com.wafflestudio.snutt2.lib.network.dto.core.ClassTimeDto
 import com.wafflestudio.snutt2.model.LectureTime
 import com.wafflestudio.snutt2.ui.SNUTTColors
